# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MenuStatusInput": {
        "properties": {
          "message": {
            "maxLength": 240,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "status": {
            "enum": [
              "operational",
              "degraded",
              "maintenance",
              "offline"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "MenuStatusInput",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VoteBody": {
        "properties": {
          "target_id": {
            "maxLength": 36,
            "minLength": 8,
            "title": "Target Id",
            "type": "string"
          },
          "target_type": {
            "pattern": "^(trusted|community)$",
            "title": "Target Type",
            "type": "string"
          },
          "value": {
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Value",
            "type": "integer"
          }
        },
        "required": [
          "target_type",
          "target_id",
          "value"
        ],
        "title": "VoteBody",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "ii Engine API",
    "version": "0.2.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/community/mods": {
      "get": {
        "operationId": "community_mods_v1_community_mods_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Community Mods"
      },
      "post": {
        "description": "Anyone can submit a community DLL for review. Approved mods appear in the public list.",
        "operationId": "create_community_mod_v1_community_mods_post",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 1,
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "description",
            "required": true,
            "schema": {
              "maxLength": 1000,
              "minLength": 1,
              "title": "Description",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "filename",
            "required": true,
            "schema": {
              "maxLength": 180,
              "minLength": 5,
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Community Mod"
      }
    },
...
    "/v1/community/mods/{identifier}/thumbnail": {
      "get": {
        "operationId": "community_mod_thumbnail_v1_community_mods__identifier__thumbnail_get",
        "parameters": [
          {
            "in": "path",
            "name": "identifier",
            "required": true,
            "schema": {
              "title": "Identifier",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Community Mod Thumbnail"
      }
    },
...
    "/v1/developer/operations/menu": {
      "post": {
        "operationId": "set_menu_status_v1_developer_operations_menu_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuStatusInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Menu Status"
      }
    },
...
    "/v1/menu/bridge/entitlements": {
      "get": {
        "description": "Menu-authenticated: Authorization: Bridge <ticket> \u2192 live Pro status.",
        "operationId": "bridge_entitlements_v1_menu_bridge_entitlements_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bridge Entitlements"
      }
    },
    "/v1/menu/bridge/ticket": {
      "post": {
        "description": "Desktop-authenticated: mint a bridge ticket + file payload for iisStupidMenu/.",
        "operationId": "issue_bridge_ticket_v1_menu_bridge_ticket_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Issue Bridge Ticket"
      }
    },
    "/v1/mods/vote": {
      "post": {
        "operationId": "cast_vote_v1_mods_vote_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cast Vote"
      }
    },
    "/v1/mods/weekly-picks": {
      "get": {
        "description": "Top 3 mods by net score over the last 7 days (trusted + community).",
        "operationId": "weekly_picks_v1_mods_weekly_picks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Weekly Picks"
      }
    },
...
    "/v1/trusted-mods/{identifier}/thumbnail": {
      "get": {
        "operationId": "trusted_mod_thumbnail_v1_trusted_mods__identifier__thumbnail_get",
        "parameters": [
          {
            "in": "path",
            "name": "identifier",
            "required": true,
            "schema": {
              "title": "Identifier",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Trusted Mod Thumbnail"
      }
    },
...
  }
}